+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - Added ICD.Connect.Themes.Crosspoints project with XP3 UI Bindings
 
 ## [6.3.0] - 2022-07-01
 ### Changed